# Use `AhoCorasick` to speed up quote match