@charliermarsh 2024-02-02
I bet we could speed this up even more but this feels like good bang for buck: https://lemire.me/blog/2023/07/14/recognizing-string-prefixes-with-simd-instructions/